mon/OSDMonitor: have `create` behavior on `new`
`osd new` will eventually replace `osd create`, and so we want to keep
as close of a behavior as possible. Some concessions needed to be made
however, specifically in terms of what the command arguments mean for
one command and the other.

For instance, and we think this is possibly the most contentious
decision, while specifying the `id` is absolutely optional, once it has
been specified we will take one of the following approaches:

 1. the osd is destroyed, and as such we will recreate the replacement
    osd.

 2. the osd exists and is not destroyed, thus we will assume it is an
    attempt at an idempotent operation; if not idempotent, fail because
    we then assume the osd was meant to be destroyed but it is not.

A commit should follow this patch series documenting the new commands.

Signed-off-by: Joao Eduardo Luis <joao@suse.de>