ceph-disk: use osd new instead of osd create
When using --dmcrypt, the lockbox stores the OSD id and the cephx secret
that will be used when activating the OSD.

When activating, the OSD id is copied from the lockbox if available,
otherwise it is obtained from osd new.

Add support for re-using an OSD id via the --osd-id option to prepare.

Signed-off-by: Loic Dachary <loic@dachary.org>