mgr: fix MgrStandby eating messages
It was failing to pass up the Mgr::ms_dispatch
return value, so messages were being consumed
even if Mgr was returning false.  Mgr returns
false to enable Client to see FSMap, so the
symptom was the Client not working.

While we're here, refactor switch into if, because
there were only two paths in the switch and we
don't expect to be adding more.

Signed-off-by: John Spray <john.spray@redhat.com>