mgr: Zabbix monitoring module
This ceph-mgr module will pull various values from the Ceph cluster
and send them to a Zabbix Server using zabbix_sender.

This requires the zabbix_sender executable to be present on the system
running ceph-mgr as it will be invoked to send data to Zabbix.

A Zabbix template can be found in this directory which can be used
to easily get data from your Ceph cluster into Zabbix.

More information is available in the README file found in the module's
directory.

Signed-off-by: Wido den Hollander <wido@42on.com>