rgw: adminops API now supports tenant param for user creation
Allow `tenant` as a param for user creation API, also document this.
Currently we still return a -ENOENT when an invalid tenant name is
specified, while we could make it return -ERR_INVALID_TENANT, this would
make rgw admin cli not return -ENOENT when an invalid tenant name is
specified.

Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>