mon: extensible output format for health checks
The summary field can be extended with the fields
used to construct the message (e.g. including
the down osd count in the message about osds
being down).

The detail entries, similarly, can be extended
with machine-readable fields like the PG ID
for a damaged PG.

For the moment all the internal stuff is just
strings still, but we change the output format
so that we don't break it later when we
add things.

Signed-off-by: John Spray <john.spray@redhat.com>