librbd: Initialization of state inside module
Fixes the coverity issues:

** 1351715 Uninitialized scalar field
5. uninit_member: Non-static class member m_state is not initialized in this
constructor nor in any functions that it calls.
7. uninit_member: Non-static class member m_copyup_start is not initialized
in this constructor nor in any functions that it calls.
CID 1351715 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
9. uninit_member: Non-static class member m_copyup_end is not initialized
in this constructor nor in any functions that it calls.

** 1351724 Uninitialized scalar field
CID 1351724 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_state is not initialized in
this constructor nor in any functions that it calls.

** 1351725 Uninitialized scalar field
CID 1351725 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_state is not initialized in
this constructor nor in any functions that it calls.

** 1351726 Uninitialized scalar field
CID 1351726 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
3. uninit_member: Non-static class member m_state is not initialized in
this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com