rados: Initializing members of librados/TestCase.h
Fixes the coverity issues:

** 1409698 Uninitialized pointer field
2. uninit_member: Non-static class member cluster is not initialized in
this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member ioctx is not initialized in
this constructor nor in any functions that it calls.
CID 1409698 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
6. uninit_member: Non-static class member alignment is not initialized
in this constructor nor in any functions that it calls.

** 1409699 Uninitialized pointer field
2. uninit_member: Non-static class member cluster is not initialized in
 this constructor nor in any functions that it calls.
CID 1409699 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
4. uninit_member: Non-static class member ioctx is not initialized in
this constructor nor in any functions that it calls.

** 1409701 Uninitialized pointer field
2. uninit_member: Non-static class member cluster is not initialized in
this constructor nor in any functions that it calls.
CID 1409701 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
4. uninit_member: Non-static class member ioctx is not initialized in
this constructor nor in any functions that it calls.

** 1409703 Uninitialized pointer field
2. uninit_member: Non-static class member cluster is not initialized in
this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member ioctx is not initialized in
this constructor nor in any functions that it calls.
CID 1409703 (#1 of 1): Uninitialized pointer field (UNINIT_CTOR)
6. uninit_member: Non-static class member alignment is not initialized in
this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar amitkuma@redhat.com