client: reset unmounting flag to false when starting a new mount
Once the "unmounting" flag goes to true, it never flips back to false.
We don't have a lot of checks for "unmounting" in the code, but they're
in some subtle places today and I plan to add more later.

It's not clear to me whether it's possible (or advisable) to reuse a
Client that has previously been unmounted. It most certainly won't work
as expected today, and the checks I'll be adding will break it for sure.

Make sure we clear the unmounting flag when we go to re-mount the
Client.

Signed-off-by: Jeff Layton <jlayton@redhat.com>