rgw: rgw_rados: set_attrs now sets the same time for BI & object
`RGWRados::set_attrs()` used to set slightly different mtimes for the
object and in the bucket dir entry as we do an object write and set the
time at bucket index as the time at that point. Fix this by using the
setting the same mtime for the object as well.

Fixes: http://tracker.ceph.com/issues/21200
Signed-off-by: Abhishek Lekshmanan <abhishek@suse.com>