messages: Initialization of members MMDSCacheRejion
Fixes the coverity issues:

** 717283 Uninitialized scalar field
2. uninit_member: Non-static class member nonce is not initialized
in this constructor nor in any functions that it calls.
4. uninit_member: Non-static class member caps_wanted is not initialized
in this constructor nor in any functions that it calls.
6. uninit_member: Non-static class member filelock is not initialized
in this constructor nor in any functions that it calls.
8. uninit_member: Non-static class member nestlock is not initialized
in this constructor nor in any functions that it calls.
CID 717283 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
10. uninit_member: Non-static class member dftlock is not
initialized in this constructor nor in any functions that it calls.

** 717284 Uninitialized scalar field
2. uninit_member: Non-static class member nonce is not initialized
in this constructor nor in any functions that it calls.
CID 717284 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member dir_rep is not initialized
in this constructor nor in any functions that it calls

Signed-off-by: Amit Kumar <amitkuma@redhat.com>