librbd,test: Initialize m_snap_list_ret,m_,value
Fixes the coverity issues:

** 1396105 Uninitialized scalar field
CID 1396105 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member m_snap_list_ret is not
initialized in this constructor nor in any functions that it calls.

** 1396141 Uninitialized scalar field
CID 1396141 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
2. uninit_member: Non-static class member value is not initialized
in this constructor nor in any functions that it calls.

** 1396162 Uninitialized scalar field
2. uninit_member: Non-static class member m_num is not initialized
in this constructor nor in any functions that it calls.
CID 1396162 (#1 of 1): Uninitialized scalar field (UNINIT_CTOR)
4. uninit_member: Non-static class member m_len is not initialized
in this constructor nor in any functions that it calls.

Signed-off-by: Amit Kumar <amitkuma@redhat.com>