osd: fail if unable to update device class in crush
device-class is supported since luminous, so we should fail the init()
if device-class is not updated in crush.

this reverts e53798d9: if the device class changes then the user should
either follow the instruction to "ceph osd crush rm-device-class" or fix
CrushWrapper::update_device_class().

this silences clang analyzer warning of:

	Value stored to 'r' is never read

Signed-off-by: Kefu Chai <kchai@redhat.com>