OSD: Silence unused variable warnings
Even though extending the lifetime of this temporary delays execution
of its destructor, the compiler does not consider the reference to
have a side-effecting constructor.

The compiler is wrong, but we can silence its complaints.

This will cease to be an issue once we adopt C++17, since mandatory
RVO will solve the inability to bind a variable to a returned
move-only object and the need to extend a temporary.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>