Revert "Merge PR #19369 into master"
This reverts commit 01148713ad61bb2fd96d381a6f872bb509663ec7, reversing
changes made to 9e30714a0e039cb3df03dad4a31b50ef9c2de318.

Despite the change in json format being positive, the unfortunate side-effect
is that it broke upgrade testing (because the QA framework must handle the
transition of mdsmap["info"] to a list from object) and the ceph-mgr.

Fixes: http://tracker.ceph.com/issues/22527
Signed-off-by: Patrick Donnelly <pdonnell@redhat.com>