mgr: remove old 'config set' command
This is a cli command to set a config option on the mgr.  It doesn't
really belong because it only sets it on the mgr that happens to be
active.  It was probably mixed up with the asok or tell command.

Added in 8b16ec31743c7b979894fddd2af9b69042b05694

Signed-off-by: Sage Weil <sage@redhat.com>