rbd: properly pass ceph global command line args to subprocess
When initializing the global context, global_init parses command line
arguments and removes ceph global args from the list. As a result they
are not visible for rbd command line parser and global options were
ignored when passing them to subprocesses like nbd or ggate.

Fix this by keeping a list of ceph global init command line
arguments. To build the list we compare the list of initial command
line args and those that remain after global_init parsing.

After this fix it is possible to pass any ceph global option to a
subprocess.

Signed-off-by: Mykola Golub <mgolub@suse.com>